importer/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
dotenv==0.9.9
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
python-dotenv==1.2.1
python-slugify==8.0.4
requests==2.32.5
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.2
yarl==1.22.0
